plugins/requirements.txt
boto3
requests
pyyaml
napari
npe2
numpy